+++ ML/train.py
@@ -3,6 +3,7 @@ from collections import defaultdict
 from tqdm import tqdm
 import cmath
 import os
+import shutil
 import yaml
 
 from QCPcircuit import QCPcircuit
@@ -82,9 +83,17 @@ def train(dataset,
           q_pp,
           𝓧,
           fidelity):
-          
-    with open("angles.yaml", 'w') as yaml_file:
-        yaml.dump(dict(), yaml_file)
+
+    if not os.path.exists("createdParams"): os.mkdir("createdParams")
+    param_path = os.path.join("createdParams", f"{dataset}_{syntax}-{ansatz}_{layers}_{q_s}_{q_n}_{q_pp}–{𝓧}_{fidelity}.yaml")
+    
+    if os.path.exists(param_path):
+        if os.path.exists("angles.yaml"): os.remove("angles.yaml")
+        shutil.copyfile(param_path, "angles.yaml")
+    else:
+        with open("angles.yaml", 'w') as yaml_file:
+            yaml.dump(dict(), yaml_file)
+
     circuits_path = create_circuits(dataset=dataset,
                                     syntax=syntax,
                                     ansatz=ansatz,
@@ -102,8 +111,7 @@ def train(dataset,
         print(classificator.meta, classificator.correct)
         classificator.apply_gradient_descent()
 
-    if not os.path.exists("createdparams"): os.mkdir("createdParams")
-    param_path = os.path.join("createdParams", f"{dataset}_{syntax}-{ansatz}_{layers}_{q_s}_{q_n}_{q_pp}–{𝓧}_{fidelity}.yaml")
+    if os.path.exists(param_path): os.remove(param_path)
     os.rename("angles.yaml", param_path)
 
     return param_path
